Properly calculate remaining part of $ud and $udlen in tpdu_decode_ud()
In case TPDU is truncated or damaged, we need to be careful how to
interpret remaining part of extracted UD. Add checks and ensure that we
never access space after the end of input TPDU.

This fixes parsing of TPDU with valid user header and without any other
user data (also without 7-bit padding between user header and data).
Because possible padding in this case can be non-empty, tpdu_decode_ud()
always called substr with expected padding length for user data, even in
case there is no user data (and so no padding). Which resulted in substr
warning: 'substr outside of string'.